fixed warning
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@4925 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/MSXMatsushita.cc
@@ -51,7 +51,7 @@ byte MSXMatsushita::peekIO(word port, const EmuTime& /*time*/) const
 	byte result;
 	switch (port & 0x0F) {
 	case 0:
-		result = ~ID;
+		result = static_cast<byte>(~ID);
 		break;
 	case 1:
 		result = firmwareSwitch->getStatus() ? 0x7F : 0xFF;
